Actually, there are two interfaces you can use to get the same markdown service. How do I know these will give us the same object? And which should we use? We'll talk about those two questions in the next chapter.
But since it doesn't matter, let's use MarkdownInterface. Open ArticleController. In show(), create a new variable -  - and set it to the multiline HEREDOC syntax. I'm going to paste in some fake content. This is the same beefy content that's in the template. In the controller, let's markdownify some stuff! Add some emphasis to jalapeno bacon ands let's turn beef ribs into a link to https://baconipsum.com/
And now, in the template, remove all the old stuff and just print {{ articleContent }}
Let's try it! Go back to our site and refresh! No surprise: it's the raw content. Now it's time to process this through Markdown!