Using Symfony's Cache
Let's use the AdapterInterface. Go back to our controller. Here's our next mission: to cache the markdown transformation: there's no reason to do that on every request! At the top of the method, add AdapterInterface cache.

Cool! Let's go use it! Symfony's cache service implements the PHP-standard cache interface, called PSR-6... in case you want Google it and geek-out over the details. But, you probably shouldn't care about this... it just means better interoperability between libraries. So... I guess... yay!

But... there's a downside.... a dark side. The standard is very powerful... but kinda weird to use at first. So, watch closely.

Start with {dollar_sign}item = {dollar_sign}cache->getItem(). We need to pass this a cache key. Use markdown_ and then md5({dollar_sign}articleContent)

Excellent! Different markdown content will have a different key. Now, when we call getItem() this does not actually go and fetch that from the cache. Nope, it just creates a CacheItem object in memory that can help us fetch and save to the cache.

For example, to check if this key is not already cached, use if (! {dollar_sign}item->isHit()). Inside we need to put the item into cache.
That's a two-step process.
Step 1: {dollar_sign}item->set() and then the value, which is {dollar_sign}markdown->transform({dollar_sign}articleContent).
Step 2: {dollar_sign}cache->save({dollar_sign}item)

I know, I know - it smells a bit over-engineered... but it's crazy powerful and insanely quick.

Tip

In Symfony 4.1, you will be able to use the Psr\SimpleCache\CacheInterface type-hint to get a 'simpler' (but less powerful) cache object.

After all of this, add {dollar_sign}articleContent = {dollar_sign}item->get() to fetch the value from cache.

Debugging the Cache

Ok, let's do this! Find your browser and refresh! Check this out: remember that we have a web debug toolbar icon for the cache! I'll click and open that in a new tab.

Hmm. There are a number of things called 'pools'. Pools are different cache systems and most are used internally by Symfony. The one we're using is called cache.app. And, cool! We had a cache miss and two calls: we wrote to the cache and then read from it.

Refresh the page again... and re-open the cache profiler. This time we hit the cache. Yes!