Here's our next challenge. In config/packages/framework.yaml, we configured the cache to use APCu. What if we did want to use this for production, but in the dev environment, we wanted to use the filesystem cache instead for simplicity. How could we do that?
We already know the answer! We just need to override this key inside the dev environment. Create a new file in config/packages/dev called framework.yaml... though technically, this could be called anything. We just need the same keys: framework, cache, app. Add those, but now set app to cache.adapter.filesystem, which was the original value

Let's see if it worked! Open ArticleController and dump the  object so we can see what it looks like.
And, refresh! Yes! It's using the FilesystemAdapter!