Dependency Injection: The Wrong Way First
In MarkdownHelper, oh, update the code to use the  variable. Here's the problem: MarkdownHelper needs the cache and markdown services. To say it differently, they're dependencies. So how can we get them from here?

Symfony follows object-orientated best practices... which means that there's no way to magically fetch them out of thin air. But that's no problem! If you ever need a service or some config, just pass them in.

The easiest way to do this is to add them as arguments to parse(). I'll show you a different solution in a minute - but let's get it working. Add AdapterInterface  and MarkdownInterface .

If you try it now... it fails:

    Too few arguments passed to parse(): 1 passed, 3 expected.