Let's add some logging to MarkdownHelper. As always, we just need to find which type-hint to use. Run:
./bin/console debug:autowiring

and look for log. We've seen this before: LoggerInterface. To get this in MarkdownHelper, just add a third argument: LoggerInterface . Like before, we need to create a new property and set it below. Great news! PhpStorm has a shortcut for this! With your cursor on logger, press alt+enter, select 'initialize fields' and hit Ok.

Awesome! Down in parse(), if the source contains the word bacon... then of course, we need to know about that!

Ok, try it! This article does talk about bacon. Refresh! To see if it logged, open the profiler and go to 'Logs'. Yes! Here is our message. I love autowiring.
The Other Loggers

Go back to your terminal. The debug:autowiring output say that LoggerInterface is an alias to monolog.logger. That is the id of the service that's being passed to us. Fun fact: you can get a bit more info about a service by running:

./bin/console debug:container --show-private monolog.logger

This is cool - but you could also learn a lot by dumping it. Anyways, we normally use debug:container to list all of the services in the container. But we can also get a filtered list. Let's find all services that contain the word 'log':

./bin/console debug:container --show-private log

There are about 6 services that I'm really interested in: these monolog.logger. something services.

Logging Channels

Here's what's going on. Symfony uses a library called Monolog for logging. And Monolog has a feature called channels, which are kind of like categories. Instead of having just one logger, you can have many loggers. Each has a unique name - called a channel - and each can do totally different things with their logs - like write them to different log files.

In the profiler, it even shows the channel. Apparently, the main logger uses a channel called app. But other parts of Symfony are using other channels, like request or event. If you look in config/packages/dev/monolog.yaml, you can see different behavior based on the channel. For example, most logs are saved to a dev.log file. But, thanks to this channels: ['!event'] config, which means 'not event', anything logged to the 'event' logger is not saved to this file.

This is a really cool feature. But mostly... I'm telling you about this because it's a great example of a new problem: how could we access one of these other Logger objects? I mean, when we use the LoggerInterface type-hint, it gives us the main logger. But what if we need a different Logger, like the 'event' channel logger?