Fetching a Non-Standard Service
So how can we tell Symfony to not pass us the 'main' logger, but instead to pass us the monolog.logger.markdown service? This is our first case where autowiring doesn't work.

That's no problem: when autowiring doesn't do what you want, just... correct it! Open config/services.yaml. Ignore all of the configuration on top for now. But notice that we're under a key called services. Yep, this is where we configure how our services work. At the bottom, add App\Service\MarkdownHelper, then below it, arguments.

The argument we want to configure is called logger. Use that here: logger. We are telling the container what value to pass to that argument. Use the service id: monolog.logger.markdown. Paste!

Find your browser and... try it! Bah! A big error:

    Argument 3 passed to MarkdownHelper __construct must implement LoggerInterface, string given.