Boo! How can we fix the problem from previous commit? Create a new config file called services_dev.yaml. This is the built-in way to create an environment-specific services file. And you can see that we actually started with one for the test environment. Inside, copy the code from the dev framework.yaml and paste it here.
Oh, and delete the old framework.yaml file. Now, refresh!

Woo! It works!

And that's really it! In framework.yaml, we just reference the parameter... which can be set in any other file. Like in this case: we set it in services.yaml and override it in services_dev.yaml. Actually, if you think about it, since framework.yaml is loaded first, the parameter isn't even defined at this point. But that's ok: you can reference a parameter, even if it's not set until later. Nice!