Using a Parameter in a Service
But wait, there's more! We can also use parameters inside our code - like in MarkdownParser. Suppose that we want to completely disable caching when we're in the dev environment.

How can we do that? Add a new argument called {dollar_sign}isDebug. Yep, in addition to other services, if your service has any config - like isDebug or an API key - those should also be passed as constructor arguments.

The idea is that we will configure Symfony to pass true or false based on our environment. I'll press Alt+Enter and select 'Initialize Fields' so that PhpStorm creates and sets that property for me. Below, we can say: if {dollar_sign}this->isDebug, then just return the uncached value.

Notice: this is the first time that we've had a constructor argument that is not a service. This is important: Symfony will not be able to autowire this value. Sure, we gave it a bool type-hint, but that's not enough for Symfony to guess what we want. Oh, and reverse my logic - I had it backwards!

To see that the argument cannot be autowired, refresh! Yep! A clear message:

    Cannot autowire service MarkdownHelper: argument  must have a type-hint or be given a value explicitly.