Built-in kernel.* Parameters
To set the  argument to the correct value, we could create a parameter, set it to false in services.yaml, override it in services_dev.yaml, and use it under bind.

But don't do it! Symfony already has a parameter we can use! In your terminal, the debug:container command normally lists services. But if you pass --parameters, well, you can guess what it prints:
php bin/console debug:container --parameters

Just like with services, most of these are internal values you don't care about. But, there are several that are useful: they start with kernel., like kernel.debug. That parameter is true most of the time, but is false in the prod environment.

Oh, and kernel.project_dir is also really handy. Copy kernel.debug, move back to services.yaml, and use %kernel.debug%.

Try it! Refresh! It still works!

Ok, it's time to talk a little bit more about controllers. It turns out, they're services too!