Autowiring works in exactly two places. First it works for controller actions. Arguments can either have the same name as a route wildcard - that's actually not autowiring - just a feature of controllers - or have a type-hint for a service.
Tip

Actually, there are a few other types of arguments you can get in your controller. You'll learn about them as we continue!

The second place autowiring works is the __construct() method of services. And actually, this is the real place where autowiring is meant to work: Symfony's container - and its autowiring logic - is really good at instantiating objects.

Binding Non-Service Arguments to Controllers

In services.yaml, we added an {dollar_sign}isDebug bind and used it inside of MarkdownHelper. So... could we also add an {dollar_sign}isDebug argument to a controller function? It certainly makes sense, so let's try it!

Add bool {dollar_sign}isDebug - the bool part is optional, it doesn't change any behavior. Below, just dump it.

Try it. Refresh! And... woh! It does not work!

    Controller ArticleController::show() requires that you provide a value for the  argument.

This... probably should work, and there's a good chance that it will work in the future. The bind functionality is relatively new. And it has one edge-case that does not currently work: you cannot use it to bind non-service arguments to controllers.

I know, kinda weird. But, we're working on it - and it might already work by the time you watch this. Yay open source!