I'll press Alt+Enter and select 'Initialize Fields' to create an  property and set it. I don't actually need to use this - but let's keep it as an example - I'll add a comment.
So, it's not as convenient as fetching a value via an argument to your controller action, but it works just the same. And actually, like I mentioned earlier, the container's job is really to instantiate services. And so autowiring should really only work for __construct() functions! In fact, the only reason that it also works for controller actions is for convenience! Yep, one core Symfony dev once said to another:

    Hey! Autowiring is great! But since it's so common to need services in a controller function, maybe we should make it work there too!

And then some other core developer said:
    Oh man, that sounds great! Virtual high-five!

This is not really that important. But the point is this: Symfony's container is great and instantiating service objects and using autowiring to pass values to their constructor. But every other function call - um, except for controller actions - will not have this magic. So don't expect it.

Thanks to bind, we can define what values are passed to specific argument names. But, we can go further and control what value should be passed for a specific type hint.