Let's do something fun! Google for SlackBundle - you'll find one called nexylan/slack-bundle. This is a fun library that gives us a service that can send messages to a Slack channel. To install it, find the composer require line, copy that, move over to your terminal and paste:
composer require nexylan/slack-bundle php-http/guzzle6-adapter

Interesting: this installs the bundle and some other library called guzzle6-adapter. Wait for it to install and... it fails!

Don't panic. There are two important things happening. First, this installed two bundles! Cool! You can see both of them inside bundles.php

For the SlackBundle, it says 'auto-generated recipe'. That means that the bundle doesn't actually have a recipe... but Symfony Flex, at least added it to bundles.php for us. By the way, this is not necessarily a bad thing: sometimes a bundle doesn't really need a custom recipe!

The second bundle did have a recipe. Before I started recording, I committed my changes so far. To see what that recipe did, let's run:

git status

Interesting: it added a new configuration file called httplug.yaml

We don't know what this does, but it probably configures some sensible defaults. Let's ignore it unless the docs tells us otherwise.
When composer require Fails

The second important thing I want to talk about is... well... this big error!

    The child node 'endpoint' at path 'nexy_slack' must be configured.

Oof... this is not a great error message. It means that this bundle requires some configuration, which we don't have yet. And since it didn't add a configuration file via a recipe, we'll need to create it ourselves. But before we do that, the most important thing to understand is this: when you see an error like this after running composer require, Composer did finish successfully and the library was installed.